app/requirements.txt
flask
pillow
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
tensorboard>=2.12.0